+++ src/ladb_opencutlist/ruby/constants.rb
@@ -2,6 +2,6 @@ module Ladb::OpenCutList
 
     EXTENSION_NAME = 'OpenCutList'.freeze
     EXTENSION_VERSION = '1.9.1-dev'.freeze
-    EXTENSION_BUILD = '202006140839'.freeze
+    EXTENSION_BUILD = '202006250514'.freeze
 
 end